Version 1.3.0
Hello everyone, I've been going through a lot the past year, my mental health is pretty low so it explains the lack of updates, I'm terribly sorry about it.
Though today I had a short burst of motivation to work on this mod, there are not a lot of course but it's the least I can do!

- Fixed a crash caused by missing curse particles for Toffee (fox entity)
- Updated the sprite of 20 neuros from Miyune to MinikoMew
- Added 8 new songs:
	* League of Legends and NewJeans - GODS (duet)
	* Linkin Park - The Emptiness Machine (duet)
	* Initial D - Deja Vu (neuro)
	* Icarus - STARSET (evil)
	* Paramore - Decode (evil)
	* Set It Off - Killer In The Mirror (evil)
	* KLOUD - VIRUS (evil)
	* SnowBlood - Crazy Fuckin' Robot Body (evil)

- Compressed songs to reduce the mod size, we went from 197mb (23 songs) to 183mb (31 songs)
- Added missing new legendary weapon to Teru